Don't panic if a syscall fails
Instead pass the err back as well. This function is only used for a
startup log message, so it'll be thrown away for now anyway.

Because it's only used in a log message it shouldn't necessarily panic.
For example, Getrlimit might just not be supported by the running
system, but prometheus could run fine otherwise.